@@ -14,12 +14,8 @@
     "appId": "es.urjc.bike3s",
     "win": {
       "target": [
-        {
-          "target": "nsis"
-        },
-        {
-          "target": "portable"
-        }
+        "nsis",
+        "portable"
       ],
       "icon": "icon.ico"
     },
@@ -35,15 +31,15 @@
     },
     "asar": false,
     "files": [
-      "!**/node_modules/*"
+      "!/node_modules/*"
     ]
   },
   "dependencies": {
     "@angular/animations": "^5.1.0",
     "@angular/cdk": "^5.1.0",
     "@angular/common": "^5.1.0",
-    "@angular/core": "^5.1.0",
     "@angular/compiler": "^5.1.0",
+    "@angular/core": "^5.1.0",
     "@angular/flex-layout": "^5.0.0-beta.14",
     "@angular/forms": "^5.1.0",
     "@angular/http": "^5.1.0",
@@ -79,7 +75,10 @@
     "moment": "^2.20.1",
     "ng-font-awesome": "github:tao-cumplido/ng-font-awesome",
     "popper.js": "^1.14.3",
+    "request": "^2.87.0",
+    "request-progress": "^3.0.0",
     "rxjs": "^5.5.5",
+    "tree-kill": "^1.2.0",
     "zone.js": "^0.8.18"
   },
   "devDependencies": {
@@ -94,7 +93,7 @@
     "@types/node": "^8.5.1",
     "@types/reflect-metadata": "0.0.5",
     "angular-ide": "^0.9.39",
-    "electron": "^1.8.1",
+    "electron": "^2.0.4",
     "electron-builder": "^20.17.1",
     "electron-installer-squirrel-windows": "^1.3.0",
     "express": "^4.16.2",
@@ -104,6 +103,7 @@
     "napa": "^3.0.0",
     "ng2-fused": "^0.5.1",
     "tslint": "^5.8.0",
-    "typescript": "^2.6.2"
+    "typescript": "^2.6.2",
+    "uglify-js": "^3.4.3"
   }
 }
