@@ -1,17 +1,19 @@
-import {Component, Inject, ViewChild} from "@angular/core";
+import {Component, Inject, ViewChild, TemplateRef} from "@angular/core";
 import {AjaxProtocol} from "../../ajax/AjaxProtocol";
 import {Layer, Rectangle, FeatureGroup, Circle, Marker} from "leaflet";
 import {LeafletDrawFunctions, FormJsonSchema, EntryPoint, Station} from "./config-definitions";
 import {ConfigurationLeaflethandler} from "./configuration.leaflethandler";
-import {NgbModal} from "@ng-bootstrap/ng-bootstrap";
-import {EntryPointDataType} from "../../../shared/configuration";
+import {NgbModal, NgbModalRef} from "@ng-bootstrap/ng-bootstrap";
+import {EntryPointDataType, GlobalConfiguration, ConfigurationFile} from "../../../shared/ConfigurationInterfaces";
 import {ConfigurationUtils} from "./configuration.utils";
-const  {dialog} = (window as any).require('electron').remote;
-import {SchemaformComponent} from "../schemaform-component/schemaform.component";
 import {ConfigurationSaveComponent} from "../configuration-save-component/configurationsave.component";
-import * as $ from "jquery";
+import { ConfigurationLoadComponent } from "../configuration-load-globalconfig/configuration-load.component";
+import { JsonTreeViewComponent } from "../jsoneditor-component/jsoneditor.component";
+import { SchemaFormGlobalComponent } from "../schemaform-global-component/schemaform-global.component";
+import { ConfDownMapComponent } from "../configuration-download-map/configuration-download-map.component";
 import * as L from 'leaflet';
-import 'leaflet-draw';
+const  {dialog} = (window as any).require('electron').remote;
+
 
 @Component({
     selector: 'configuration',
@@ -40,7 +42,7 @@ export class ConfigurationComponent {
     /*
     * Variables for configuration
     */
-    globalData = {
+    globalData: GlobalConfiguration = {
         boundingBox: {
             northWest: { latitude: 0, longitude:0 },
             southEast: { latitude: 0, longitude:0 }
@@ -48,6 +50,8 @@ export class ConfigurationComponent {
     };
     entryPoints: Array<EntryPoint> = new Array<any>();
     stations: Array<Station> = new Array<any>();
+
+    actualModalOpen: NgbModalRef;
 
     /*
     * Final configurations
@@ -59,154 +63,56 @@ export class ConfigurationComponent {
         "stations": new Array<any>()
     };
 
-    @ViewChild('globalSchemaForm') gsForm: SchemaformComponent;
+    hasBoundingBox: boolean;
+
+    @ViewChild('globalSchemaForm') gsForm: SchemaFormGlobalComponent;
+
+    @ViewChild('selecEntryPointType') epSelectModalForm: TemplateRef<any>;
+    @ViewChild('entryPoint') epModalForm: TemplateRef<any>;
+    @ViewChild('stationModal') stationModalForm: TemplateRef<any>;
+
+    @ViewChild('jsonTreeEp') jsonTreeEp: JsonTreeViewComponent;
+    @ViewChild('jsonTreeStation') jsonTreeStation: JsonTreeViewComponent;
 
     /*
     * Map control variables
     */
-    hasBoundingBox = false;                                 // Bounding box exist if true
     featureGroup: FeatureGroup = new FeatureGroup();        // List of drawn element
     drawOptions: any;                                       // Draw options for leaflet-draw
     map: L.Map;                                             // showed Map
     lastCircleAdded: Circle;                                // Last circle added
     lastMarkerAdded: Marker;                                // Last Marker added
 
-    constructor(@Inject('AjaxProtocol') private ajax: AjaxProtocol,
-                private modalService: NgbModal) {}
+    constructor(@Inject('AjaxProtocol') public ajax: AjaxProtocol,
+                public modalService: NgbModal) {}
 
     async ngOnInit() {
+        this.hasBoundingBox = false;
         this.drawOptions = LeafletDrawFunctions.createLeafletDrawOptions(this.featureGroup);
         (L as any).drawLocal = LeafletDrawFunctions.createCustomMessages();
         await this.ajax.formSchema.init();
-        this.globalFormInit().then();
-        this.selectEntryPointFormInit().then();
-        this.stationFormInit().then();
-    }
-
-    getMapController(map: L.Map) {
-        this.map = map;
-        this.drawOptions.edit.featureGroup.addTo(this.map);
-        this.defineMapEventHandlers();
-    }
-
-    defineMapEventHandlers() {
-        this.map.on('draw:created', (event: any) => {
-            let layer: Layer = event.layer;
-            let type = event.layerType;
-            switch(type) {
-                case 'rectangle': ConfigurationLeaflethandler.onDrawBoundingBoxHandler(<Rectangle> layer, this);
-                break;
-                case 'circle': ConfigurationLeaflethandler.onDrawEntryPointHandler(<Circle> layer, this);
-                break;
-                case 'marker': ConfigurationLeaflethandler.onDrawStationHandler(<Marker> layer, this);
-            }
-        });
-        this.map.on('draw:deleted', (event: any) => {
-            for (let layer of event.layers.getLayers()) {
-                if(layer instanceof Rectangle && this.hasBoundingBox) {
-                    ConfigurationLeaflethandler.onDeleteBoundingBoxHandler(this);
-                }
-                if (layer instanceof Circle) {
-                    let index = ConfigurationUtils.getEntryPointIndexByCircle(layer, this.entryPoints);
-                    if (index !== null) {
-                        this.entryPoints.splice(index, 1);
-                        this.finalEntryPoints.entryPoints.splice(index, 1);
-                    }
-                    this.featureGroup.removeLayer(layer);
-                }
-                if (layer instanceof Marker) {
-                    let index = ConfigurationUtils.getStationIndexByMarker(layer, this.stations);
-                    if(index !== null) {
-                        this.stations.splice(index, 1);
-                        this.finalStations.stations.splice(index, 1);
-                    }
-                    this.featureGroup.removeLayer(layer);
-                }
-            }
-        });
-    }
-
-    globalFormSubmit($event: any) {
-        this.globalData = $event;
-        console.log('Global Configuration Updated');
-        console.log(this.globalData);
-    }
-
-    isGlobalFormValid($event: any) {
-        console.log('Is global Form Valid?: ' + $event);
-        this.globalConfigValid = $event;
-    }
-
-    selectEntryPointSubmit(data: EntryPointDataType) {
-        $('#form-select-entry-point-close').trigger('click');
-        console.log(data);
-        this.lastSelectedEntryPointType = data;
-        this.entryPointFormInit(data).then((entryPointData: any) => {
-            entryPointData.position.latitude = this.lastCircleAdded.getLatLng().lat;
-            entryPointData.position.longitude = this.lastCircleAdded.getLatLng().lng;
-            if(this.entryPointForm.schema.properties.radius !== null) {
-                entryPointData.radius = this.lastCircleAdded.getRadius();
-            }
-            $('#form-entry-point-button').trigger('click');
-        });
-    }
-
-    entryPointSubmit(entryPoint: any) {
-        entryPoint.entryPointType = {};
-        entryPoint.entryPointType = this.lastSelectedEntryPointType.entryPointType;
-        entryPoint.userType.typeName = this.lastSelectedEntryPointType.userType;
-        this.lastCircleAdded.setLatLng({
-            lat: entryPoint.position.latitude,
-            lng: entryPoint.position.longitude
-        });
-        if(this.lastSelectedEntryPointType.entryPointType === "POISSON") {
-            this.lastCircleAdded.setRadius(entryPoint.radius);
-        } else {
-            this.lastCircleAdded.setRadius(50);
-        }
-        let newEntryPoint = new EntryPoint(entryPoint, this.lastCircleAdded);
-        newEntryPoint.getCircle().bindPopup(newEntryPoint.getPopUp());
-        this.entryPoints.push(newEntryPoint);
-        this.finalEntryPoints.entryPoints.push(newEntryPoint.getEntryPointInfo());
-        console.log(this.entryPoints);
-        $('#form-entry-point-close').trigger('click');
-    }
-
-    stationSubmit(station: any) {
-        this.lastMarkerAdded.setLatLng({
-            lat: station.position.latitude,
-            lng: station.position.longitude
-        });
-        let newStation = new Station(station, this.lastMarkerAdded);
-        this.stations.push(newStation);
-        this.finalStations.stations.push(newStation.getStationInfo());
-        this.lastMarkerAdded.setIcon(newStation.getIcon());
-        $('#form-station-close').trigger('click');
+        this.globalFormInit();
+        this.selectEntryPointFormInit();
+        this.stationFormInit();
     }
 
     async globalFormInit() {
-        this.ajax.formSchema.getGlobalSchema().then((data) => {
-            console.log(data);
-            this.globalForm = {
-                schema: JSON.parse(data),
-                data: this.globalData,
-                options: {
-                    addSubmit: false
-                }
-            };
-            console.log(this.globalForm);
-            return;
-        });
+        let schema = await this.ajax.formSchema.getGlobalSchema();
+        this.globalForm = {
+            schema: JSON.parse(schema),
+            data: this.globalData,
+            options: {
+                addSubmit: false
+            }
+        };
     }
 
     async selectEntryPointFormInit(): Promise<void> {
-        this.ajax.formSchema.getSchemaFormEntryPointAndUserTypes().then((schema) => {
-            this.selectEntryPointForm = {
-                schema:schema,
-                data: null
-            };
-            return;
-        });
+        let schema = await this.ajax.formSchema.getSchemaFormEntryPointAndUserTypes();
+        this.selectEntryPointForm = {
+            schema:schema,
+            data: null
+        };
     }
 
     async entryPointFormInit(selected: EntryPointDataType): Promise<any> {
@@ -223,45 +129,149 @@ export class ConfigurationComponent {
     }
 
     async stationFormInit(): Promise<void> {
-        this.ajax.formSchema.getStationSchema().then((schema) => {
-            this.stationForm = {
-                schema: JSON.parse(schema),
-                data: this.lastStation
-            };
-            return;
+        let schema = await this.ajax.formSchema.getStationSchema();
+        this.stationForm = {
+            schema: JSON.parse(schema),
+            data: this.lastStation
+        };
+    }
+
+    getMapController(map: L.Map) {
+        this.map = map;
+        this.drawOptions.edit.featureGroup.addTo(this.map);
+        this.defineMapEventHandlers();
+        console.log(this.featureGroup);
+    }
+
+    bboxChanged(data: any) {
+        Object.assign(this.globalData.boundingBox, data);
+    }
+
+    defineMapEventHandlers() {
+        this.map.on('draw:created', (event: any) => {
+            let layer: Layer = event.layer;
+            let type = event.layerType;
+            switch(type) {
+                case 'rectangle': ConfigurationLeaflethandler.drawBoundingBox(this, <Rectangle> layer);
+                break;
+                case 'circle': ConfigurationLeaflethandler.drawEntryPoint(this, <Circle> layer);
+                break;
+                case 'marker': ConfigurationLeaflethandler.drawStation(this, <Marker> layer);
+            }
+        });
+        this.map.on('draw:deleted', (event: any) => {
+            for (let layer of event.layers.getLayers()) {
+                if(layer instanceof Rectangle) {
+                    ConfigurationLeaflethandler.deleteBoundingBox(this);
+                }
+                if (layer instanceof Circle) {
+                    let index = ConfigurationUtils.getEntryPointIndexByCircle(layer, this.entryPoints);
+                    if (index !== null) {
+                        this.entryPoints.splice(index, 1);
+                        this.finalEntryPoints.entryPoints.splice(index, 1);
+                        this.jsonTreeEp.dataUpdated(this.finalEntryPoints);
+                    }
+                    this.featureGroup.removeLayer(layer);
+                }
+                if (layer instanceof Marker) {
+                    let index = ConfigurationUtils.getStationIndexByMarker(layer, this.stations);
+                    if(index !== null) {
+                        this.stations.splice(index, 1);
+                        this.finalStations.stations.splice(index, 1);
+                        this.jsonTreeStation.dataUpdated(this.finalStations);
+                    }
+                    this.featureGroup.removeLayer(layer);
+                }
+            }
         });
     }
 
-    openFormEntryPoint(modalRef: any) {
-        this.modalService.open(modalRef, {backdrop: "static", keyboard: false}).result.then((result) => {
-            if(result === 'Close click') {
-                console.log("Removing circle");
-                this.featureGroup.removeLayer(this.lastCircleAdded);
-            }
-        }, () => {});
+    globalFormSubmit($event: any) {
+        this.globalData = $event;
+        console.log(this.globalData);
+    }
+
+    isGlobalFormValid($event: any) {
+        this.globalConfigValid = $event;
+        console.log(this.globalConfigValid);
+    }
+
+    async selectEntryPointSubmit(data: EntryPointDataType) {
+        this.actualModalOpen.close();
+        console.log(data);
+        this.lastSelectedEntryPointType = data;
+        let entryPointData = await this.entryPointFormInit(data);
+        entryPointData.position.latitude = this.lastCircleAdded.getLatLng().lat;
+        entryPointData.position.longitude = this.lastCircleAdded.getLatLng().lng;
+        if(this.entryPointForm.schema.properties.radius !== null) {
+            entryPointData.radius = this.lastCircleAdded.getRadius();
+        }
+        this.actualModalOpen = this.modalService.open(this.epModalForm);
+    }
+
+    entryPointSubmit(entryPoint: any) {
+        entryPoint.entryPointType = {};
+        entryPoint.entryPointType = this.lastSelectedEntryPointType.entryPointType;
+        entryPoint.userType.typeName = this.lastSelectedEntryPointType.userType;
+        this.lastCircleAdded.setLatLng({
+            lat: entryPoint.position.latitude,
+            lng: entryPoint.position.longitude
+        });
+        if(entryPoint.hasOwnProperty('radius')) {
+            this.lastCircleAdded.setRadius(entryPoint.radius);
+        } else {
+            this.lastCircleAdded.setRadius(50);
+        }
+        let newEntryPoint = new EntryPoint(entryPoint, this.lastCircleAdded);
+        newEntryPoint.getCircle().bindPopup(newEntryPoint.getPopUp());
+        this.entryPoints.push(newEntryPoint);
+        this.finalEntryPoints.entryPoints.push(newEntryPoint.getEntryPointInfo());
+        console.log(this.entryPoints);
+        this.actualModalOpen.close();
+        if(this.jsonTreeEp) {
+            this.jsonTreeEp.dataUpdated(this.finalEntryPoints);   
+        }
+    }
+
+    stationSubmit(station: any) {
+        this.lastMarkerAdded.setLatLng({
+            lat: station.position.latitude,
+            lng: station.position.longitude
+        });
+        let newStation = new Station(station, this.lastMarkerAdded);
+        this.stations.push(newStation);
+        this.finalStations.stations.push(newStation.getStationInfo());
+        this.lastMarkerAdded.setIcon(newStation.getIcon());
+        this.actualModalOpen.close();
+        if(this.jsonTreeStation) {
+            this.jsonTreeStation.dataUpdated(this.finalStations);
+        }
+    }
+    openForm(modalRef: any) {
+        this.actualModalOpen = this.modalService.open(modalRef, {backdrop: "static", keyboard: false});
+    }
+
+    closeFormEntryPoint() {
+        this.actualModalOpen.close();
+        this.featureGroup.removeLayer(this.lastCircleAdded);
     }
 
     openStationForm(modalRef: any) {
-        this.modalService.open(modalRef, {backdrop: "static", keyboard: false}).result.then((result) => {
-            if(result === 'Close click') {
-                console.log("Removing circle");
-                this.featureGroup.removeLayer(this.lastMarkerAdded);
-            }
-        }, () => {});
+        this.actualModalOpen = this.modalService.open(modalRef, {backdrop: "static", keyboard: false});
+        this.featureGroup.removeLayer(this.lastMarkerAdded);
     }
 
-    updateGlobalFormView() {
-        let tab1 = document.getElementById('added-entities');
-        let tab2 = document.getElementById('global-form');
-        console.log("tpm");
-        if(tab1 !== null && tab2 !== null) {
-            tab1.click();
-            tab2.click();
-        }
+    closeStationForm() {
+        this.actualModalOpen.close();
+        this.featureGroup.removeLayer(this.lastMarkerAdded);
     }
 
     generateConfiguration() {
         let path = this.selectFolder();
+        if(this.isGlobalFormValid) {
+            this.globalData = this.gsForm.actualData;
+            console.log(this.globalData);
+        }
         const modalRef = this.modalService.open(ConfigurationSaveComponent);
         modalRef.componentInstance.path = path;
         modalRef.componentInstance.globalConfiguration = this.globalData;
@@ -271,7 +281,71 @@ export class ConfigurationComponent {
     }
 
     selectFolder(): string {
-        return dialog.showOpenDialog({properties: ['openDirectory', 'createDirectory']})[0];
+            return dialog.showOpenDialog({properties: ['openDirectory', 'createDirectory']})[0];
+    }
+
+    saveFile(): string {
+        return dialog.showSaveDialog({filters: [{name: 'OSM Files', extensions: ['osm']}]});
+    }
+
+    selectFile(): string | undefined {
+        return dialog.showOpenDialog({
+            properties: ['openFile', 'createDirectory'],
+            filters: [{name: 'JSON Files', extensions: ['json']}]
+        })[0].replace(/\\/g, "/");
+    }
+
+    async loadGlobalConfig() {
+        let file = this.selectFile();
+        let modalRef: NgbModalRef = this.modalService.open(ConfigurationLoadComponent);
+        modalRef.componentInstance.path = file;
+        modalRef.componentInstance.configurationFile = ConfigurationFile.GLOBAL_CONFIGURATION;
+        modalRef.result.then((globalData) => {
+            Object.assign(this.globalData, globalData);
+            this.gsForm.resetForm();
+            ConfigurationLeaflethandler.drawBoundingBox(this, this.globalData.boundingBox);
+        });
+        this.hasBoundingBox = true;
+    }
+
+    async loadEntryPoints() {
+        let file = this.selectFile();
+        let modalRef: NgbModalRef = this.modalService.open(ConfigurationLoadComponent);
+        modalRef.componentInstance.path = file;
+        modalRef.componentInstance.configurationFile = ConfigurationFile.ENTRYPOINT_CONFIGURATION;
+        modalRef.result.then((entryPointsConfig: any) => {
+            for(let entryPoint of entryPointsConfig.entryPoints) {
+                ConfigurationLeaflethandler.drawEntryPoint(this, entryPoint);
+            }
+            if(this.jsonTreeEp) {
+                this.jsonTreeEp.dataUpdated(this.finalEntryPoints);
+            }
+        });
+    }
+
+    async loadStations() {
+        let file = this.selectFile();
+        let modalRef: NgbModalRef = this.modalService.open(ConfigurationLoadComponent);
+        modalRef.componentInstance.path = file;
+        modalRef.componentInstance.configurationFile = ConfigurationFile.STATION_CONFIGURATION;
+        modalRef.result.then((stationsConfig: any) => {
+            for(let station of stationsConfig.stations) {
+                ConfigurationLeaflethandler.drawStation(this, station);
+            }
+            if(this.jsonTreeStation) {
+                this.jsonTreeStation.dataUpdated(this.finalStations);
+            }
+        });
+    }
+
+    async downloadMap() {
+        let path = this.saveFile();
+        Object.assign(this.globalData, this.gsForm.actualData.boundingBox);
+        if(path) {
+            let modalRef: NgbModalRef = this.modalService.open(ConfDownMapComponent, {backdrop: "static", keyboard: false});
+            modalRef.componentInstance.path = path;
+            modalRef.componentInstance.boundingBox = this.globalData.boundingBox;
+        }
     }
 
 }
