@@ -1,51 +1,43 @@
-import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {Component, EventEmitter, Input, OnInit, Output, ChangeDetectorRef} from '@angular/core';
 
 @Component({
     selector: 'schema-form',
     template: require('./schemaform.component.html'),
-    styles: [require('./schemaform.component.css')]
+    styles: []
 })
-export class SchemaformComponent implements OnInit {
-
+export class SchemaFormComponent {
+    
     @Input()
     form: any;
 
     @Output('dataSubmited')
     dataSubmited = new EventEmitter<any>();
 
-    @Output('automaticSubmit')
-    automaticSubmit = new EventEmitter<any>();
-
     @Output('isValid')
     isValid = new EventEmitter<any>();
 
+    reloading: boolean;
+    globalUpdate: boolean;
+
     actualData: any;
 
-    constructor() {
-    }
-
-    ngOnInit(): void {
-    }
+    constructor(private cdRef: ChangeDetectorRef) {}
 
     resetForm() {
-        console.log(this.form);
+        this.reloading = true;
+        this.cdRef.detectChanges();
+        let jsonForm = document.getElementById("json-form");
+        this.reloading = false;
+        if(jsonForm) {
+            jsonForm.click();
+        }
     }
 
     submit(data: any) {
-        console.log(data);
         this.dataSubmited.emit(data);
     }
 
     valid(isValid: any) {
         this.isValid.emit(isValid);
-        if(isValid) {
-            this.automaticSubmit.emit(this.actualData);
-        }
     }
-
-    changed(data: any) {
-        this.actualData = data;
-        console.log(data);
-    }
-
 }
