@@ -1,7 +1,7 @@
 process.env.SPARKY_LOG = false;
 
 const { FuseBox, Sparky } = require('fuse-box');
-const { EnvPlugin, CSSPlugin, CSSResourcePlugin, RawPlugin, WebIndexPlugin, JSONPlugin } = require('fuse-box');
+const { EnvPlugin, CSSPlugin, CSSResourcePlugin, RawPlugin, WebIndexPlugin, JSONPlugin, QuantumPlugin } = require('fuse-box');
 
 const log = require('fliplog');
 const express = require('express');
@@ -171,7 +171,7 @@ Sparky.task('build:data-analyser', () => {
         experimentalFeatures: true
     });
 
-    const main = fuse.bundle('data-analyser.js').instructions('> [main/DataAnalyserTool.ts]');
+    const main = fuse.bundle('data-analyser.js').instructions('>main/DataAnalyserTool.ts');
 
     return fuse.run();
 });
@@ -182,14 +182,16 @@ Sparky.task('build:frontend:main', () => {
         output: path.join(projectRoot.build.frontend(), '$name.js'),
         target: 'server',
         experimentalFeatures: true,
-        cache: !production,
+        ignoreModules: ['electron'],
         plugins: [
-            EnvPlugin({ target: production ? 'production' : 'development' })
+            EnvPlugin({ target: production ? 'production' : 'development' }),
+            JSONPlugin()
         ]
     });
 
-    const main = fuse.bundle('main').instructions('> [main/main.ts]');
+    const main = fuse.bundle('main').instructions('>main/main.ts');
 
+    
     if (!production) {
         // main.watch('main/**');
         return fuse.run().then(() => {
@@ -198,10 +200,10 @@ Sparky.task('build:frontend:main', () => {
                 shell: true, // necessary on windows
                 stdio: 'inherit' // pipe to calling process
             });
-    });
-}
+        });
+    }
 
-return fuse.run();
+    return fuse.run();
 });
 
 Sparky.task('build:frontend:renderer', () => {
@@ -223,6 +225,14 @@ Sparky.task('build:frontend:renderer', () => {
             WebIndexPlugin({
                 template: path.join(projectRoot.frontend.renderer(), 'index.html'),
                 path: '.'
+            }),
+            JSONPlugin(),
+            production && QuantumPlugin({
+                bakeApiIntoBundle : false,
+                target : 'electron',
+                treeshake: true,
+                removeExportsInterop: false,
+                uglify: true
             })
         ]
     });
@@ -249,7 +259,7 @@ Sparky.task('build:frontend:renderer', () => {
     const destination = path.join(projectRoot.build.frontend(), 'styles.css');
     fs.copySync(globalCss, destination);
 
-    const packageProdOrig = path.join(projectRoot(), 'package_production.json');
+    const packageProdOrig = path.join(projectRoot(), 'package_prod.json');
     const packageProdDest = path.join(projectRoot.build(), 'package.json');
     fs.copySync(packageProdOrig, packageProdDest);
 
